azure-functions/requirements.txt
azure-functions
azure-identity
msgraph-core
requests
